Fix build issues after upgrading KataGo to v1.16.3
- Updated CMakeLists.txt-macos to include the -Wno-c++17-extensions flag.
- Changed variable types from int to size_t in MetalProcess class for consistency.
- Implemented new merging functions for batch normalization scales and biases in SWBatchNormLayerDesc.
- Refactored tests to use merged scale and bias arrays instead of separate pointers.

+++ cpp/neuralnet/metalbackend.cpp
@@ -898,7 +898,7 @@ void MetalProcess::processScoreValues(
   const float* currentScoreValueData = &inputBuffers->scoreValuesResults[offset];
 
   if(modelVersion >= 9) {
-    int numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
+    size_t numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
     assert(numScoreValueChannels == 6);
     currentOutput->whiteScoreMean = currentScoreValueData[0];
     currentOutput->whiteScoreMeanSq = currentScoreValueData[1];
@@ -908,7 +908,7 @@ void MetalProcess::processScoreValues(
     currentOutput->shorttermScoreError = currentScoreValueData[5];
   }
   else if(modelVersion >= 8) {
-    int numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
+    size_t numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
     assert(numScoreValueChannels == 4);
     currentOutput->whiteScoreMean = currentScoreValueData[0];
     currentOutput->whiteScoreMeanSq = currentScoreValueData[1];
@@ -918,7 +918,7 @@ void MetalProcess::processScoreValues(
     currentOutput->shorttermScoreError = 0;
   }
   else if(modelVersion >= 4) {
-    int numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
+    size_t numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
     assert(numScoreValueChannels == 2);
     currentOutput->whiteScoreMean = currentScoreValueData[0];
     currentOutput->whiteScoreMeanSq = currentScoreValueData[1];
@@ -929,7 +929,7 @@ void MetalProcess::processScoreValues(
   }
   else {
     assert(modelVersion >= 3);
-    int numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
+    size_t numScoreValueChannels = inputBuffers->singleNnScoreValuesResultElts;
     assert(numScoreValueChannels == 1);
     currentOutput->whiteScoreMean = currentScoreValueData[0];
     //Version 3 neural nets don't have any second moment currentOutput, implicitly already folding it in, so we just use the mean squared
